guideline/pyproject.toml
[tool.poetry]
name = "dacon-hansoldeco-season3-guideline"
version = "0.1.0"
description = "DACON - 건설공사 사고 예방 및 대응책 생성 : 한솔데코 시즌3 AI 경진대회"
authors = ["Jeongwon <banb3515@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.20"
langchain-community = "^0.3.19"
langchain-huggingface = "^0.1.2"
python-dotenv = "^1.0.1"
docling = "^2.26.0"
tqdm = "^4.67.1"
pypdf = "^5.3.1"
pandas = "^2.2.3"
faiss-cpu = "^1.10.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
